software: Add error code in update-fpga: file doesn't exist

+++ _install/r_pi/scripts/update-fpga.sh
@@ -8,6 +8,10 @@ else
 	rfile=$1
 fi
 
+if [ ! -f ${rfile} ]; then
+	exit 1
+fi
+
 flash_erase ${mdev} 0 6
 dd if=${rfile} of=${mdev}
 
@@ -16,5 +20,5 @@ if [ $? -eq 0 ]; then
 	$(dirname $0)/fpga-reconf.sh
 	exit 0
 else
-	exit 1
+	exit 2
 fi
